closes #171
