+++ client/app/views/people/show.html
@@ -108,10 +108,12 @@
       <h3>Conditions</h3>
       <p class="explantory-text">A <em>condition</em>is like a status. At the moment you can either be <em>Active</em>, <em>Retired</em>, <em>Maternity</em>, <em>Leave</em>, <em>Deceased</em>, <em>Applicant</em>, <em>Left</em> or </em>Unknown</em>, but we'll add more as we work out what is needed. The system calculates your current condition by looking at the most recent start date, so it is important to update new conditins as soon as you can. </p>
 
-      <editable-group object="person" lens="conditions" subgroup-class="alert alert-info col-sm-6 col-md-4">
-        <editable property="name"       label="name"                  >This is the label of this period of your life, it will normally be taken from the list above.</editable>
-        <editable property="start_date" label="start date" type="date">This is the date that this period started. YOu don't need to put in when it ended becasue that is inferred from the start date of the next period.</editable>
-      </editable-group>
+      <div ng-repeat="condition in person.conditions">
+        <editable-group object="person" lens="conditions[{{condition|id}}]" subgroup-class="alert alert-info col-sm-6 col-md-4">
+          <editable property="name"       label="name"                  >This is the label of this period of your life, it will normally be taken from the list above.</editable>
+          <editable property="start_date" label="start date" type="date">This is the date that this period started. YOu don't need to put in when it ended becasue that is inferred from the start date of the next period.</editable>
+        </editable-group>
+      </div>
 
     </section>
 
@@ -129,29 +131,31 @@
         <div class="col-md-8">
           <div class="row">
           <h4>Event Participation</h4>
+          <div ng-repeat="ep in person.employee.office_culture.event_participations">
             <editable-group object="person"
-                            lens="employee.office_culture.event_participations"
+                            lens="employee.office_culture.event_participations[{{ep|id}}]"
                             subgroup-class="alert alert-info col-sm-6">
               <editable property="name" label="name"            >This is the name of the thing that you did.</editable>
               <editable property="date" label="date" type="date">This is the date that you did it.</editable>
               <editable property="url"  label="url"  type="url" >This is a link to some kind of explanation of the event. It could be the event's website, or it might be to some report that you've published about it.</editable>
               <p><a href="{{ subobject.url }}"><span class="glyphicon glyphicon-link"></span>Go to link.</a></div>
             </editable-group>
+          </div>
       </div>
     </section>
 
     <div class="row">
       <div class="col-md-6">
-        <section ng-show="isCurrentUser(person) || false && person.employee.contact">
+        <section ng-show="isCurrentUser(person) || person.employee.contact">
           <div class="row">
             <h3>contact (office)</h3>
             <p class="explantory-text">Explanatory text goes in here</p>
             <editable-group object="person"
-                            lens="employee.office_culture.event_participations"
+                            lens="employee.contact"
                             subgroup-class="alert alert-info col-sm-6">
               <editable  property="extension" label="extension"  type="number">This is your desk phone extension.</editable>
               <editable  property="studio"    label="studio"                  >This is the studio that you work in.</editable>
-              <editable  property="work email"label="work email" type="email" >This is your ~~~@bvn.com.au email address.</editable>
+              <editable  property="work_email" label="work email" type="email" >This is your ~~~@bvn.com.au email address.</editable>
             </editable-group>
             <div class="col-md-6">
               <h4>Mobile</h4>
@@ -172,8 +176,12 @@
             <h3>contact (personal)</h3>
             <div class="col-md-6">
               <h4>anti-social</h4>
-              <editable lens="contact.personal_email" label="personal email" type="email">This is</editable>
-              <editable lens="contact.personal_phone" label="personal phone" type="tel">This is</editable>
+              <editable-group object="person"
+                              lens="contact"
+                              subgroup-class="alert alert-info col-sm-6">
+                <editable lens="contact.personal_email" label="personal email" type="email">This is</editable>
+                <editable lens="contact.personal_phone" label="personal phone" type="tel">This is</editable>
+              </editable-group>
             </div>
             <div class="col-md-6">
               <h4>social</h4>
